% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigNetworkSearch.R
\name{bigNetworkSearch}
\alias{bigNetworkSearch}
\title{For a big network, apply treatments to a subset and search}
\usage{
bigNetworkSearch(A, treatments, sampleSize, restartProb = 0, numIterations)
}
\arguments{
\item{A}{adjacency Matrix of the network}

\item{treatments}{number of treatments to search over}

\item{sampleSize}{Number of nodes given a treatment, as opposed to null treatment}

\item{restartProb}{The probability of starting from a random restart. Set to 1 to evaluate random
balanced designs.}

\item{numIterations}{How many iterations to use in the exchange algorithm}
}
\description{
For a big network where we can apply treatments to a subset of them, find the optimal
design using a Fedorov exchange type algorithm.
}
