% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateNetworkDesign.R
\name{evaluateNetworkDesign}
\alias{evaluateNetworkDesign}
\title{Calculate the optimality criteria function for a given network design}
\usage{
evaluateNetworkDesign(infMatrix, p)
}
\arguments{
\item{infMatrix}{An information Matrix}

\item{p}{number of treatmentsp}
}
\description{
Calculate the optimality criteria (A_s optimality) for a given network design
}
