% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch.R
\name{gridSearch}
\alias{gridSearch}
\title{Search over all possible designs to find the global optimal}
\usage{
gridSearch(
  A,
  p,
  isomorphisms = FALSE,
  specialNodes = NULL,
  ignoreLastNode = FALSE,
  algorithm = "sequential"
)
}
\arguments{
\item{A}{An adjacency Matrix}

\item{p}{number of treatments to search over}

\item{isomorphisms}{whether to ignore designs that are isomorphic to designs evaluated earlier}

\item{specialNodes}{Are there any nodes such as nodes in block designs that we don't want to apply treatments to.}

\item{ignoreLastNode}{Whether we set the last node as zero. Can be useful to satisfy constraints in double blocking structures}

\item{algorithm}{Which algorithm to use for finding which designs to evaluate; currently only "sequential" , "random" adn "ce" are implemented}
}
\description{
Evaluates all possible designs by searching over a grid to find a global optimal. Can be slow for high A>20 or high p>2
}
\keyword{For}
\keyword{a}
\keyword{adjacency}
\keyword{design}
\keyword{find}
\keyword{given}
\keyword{matrix,}
\keyword{optimal}
\keyword{p}
\keyword{the}
\keyword{treatemtnes}
\keyword{with}
