% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearch_old.R
\name{gridSearchOld}
\alias{gridSearchOld}
\title{Search over all possible designs to find the global optimal}
\usage{
gridSearchOld(
  A,
  p,
  isomorphisms = FALSE,
  specialNodes = NULL,
  ignoreLastNode = FALSE,
  algorithm = "sequential"
)
}
\arguments{
\item{A}{An adjacency matrix}

\item{p}{The number of treatments in the experiment,}

\item{isomorphisms}{Whether to ignore designs that are isomorphic to designs evaluated earlier}

\item{specialNodes}{Are there any nodes which are fixed that we don't want to apply treatments to.}

\item{ignoreLastNode}{Whether we set the last node as zero. Can be useful to satisfy constraints in double blocking structures}

\item{algorithm}{Which algorithm to use for finding which designs to evaluate; currently 'sequential', 'random' and 'ce' are implemented}
}
\description{
Evaluates all possible designs by searching over a grid to find a global optimal. Can be slow for high A>20 or high p>2
}
\keyword{For}
\keyword{a}
\keyword{adjacency}
\keyword{design}
\keyword{find}
\keyword{given}
\keyword{matrix,}
\keyword{optimal}
\keyword{p}
\keyword{the}
\keyword{treatemtnes}
\keyword{with}
