% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSearchPrior.R
\name{gridSearchPrior}
\alias{gridSearchPrior}
\title{Find the optimal design for a given network.}
\usage{
gridSearchPrior(
  A,
  p,
  isoSearch = FALSE,
  blockList = NULL,
  ignoreLastNode = FALSE,
  algorithm = "sequential",
  networkEffects = FALSE,
  weightPrior = NULL,
  viralOpt = TRUE,
  weights
)
}
\arguments{
\item{A}{An adjacency matrix}

\item{p}{The number of treatments in the experiment}

\item{isoSearch}{Whether to ignore designs that are isomorphic to designs evaluated earlier}

\item{blockList}{A list. each element of which is a collection of nodes which form a block.}

\item{ignoreLastNode}{Whether we set the last node as zero. Set to true to satisfy constraints in double blocking structures}

\item{algorithm}{Which algorithm to use for finding which designs to evaluate; currently 'sequential', 'random' and 'ce' and exchange are implemented}

\item{networkEffects}{Set to true if the optimal design for the network effects be found (default is to find direct effects)}

\item{viralOpt}{If we have a viral parameter, do we want to estimate it (TRUE) or just estimate the difference in treatment effects}
}
\description{
Finds the optimal design for a given network. Various algorithms are implemented to search a network to find the optimal design for.
estimating treatment effects on that network. It can also be used to find optimal designs for experiments that contain blocking.
Can be slow for high A>20 or high p>2.
}
\keyword{For}
\keyword{a}
\keyword{adjacency}
\keyword{design}
\keyword{find}
\keyword{given}
\keyword{matrix,}
\keyword{optimal}
\keyword{p}
\keyword{the}
\keyword{treatments.}
\keyword{with}
