% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informationMatFull.R
\name{informationMatFull}
\alias{informationMatFull}
\title{Calculate Information Matrix}
\usage{
informationMatFull(A, X)
}
\arguments{
\item{A}{An adjacency Matrix}

\item{X}{design matrix}
}
\description{
Calculates the information matrix for a given adjacency matrix and design.
}
\keyword{Calculate}
\keyword{an}
\keyword{information}
\keyword{matrix}
