% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{JRSSExamples}
\alias{JRSSExamples}
\title{Example Networks For Demonstrating Experimental Design}
\format{
Examples of Adjacency Matrices.
\describe{
  \item{JRSSExamples}{A list containing all the Examples for simple networks from the JRSSC Paper}
  #'  ...
}
}
\source{
\url{https://doi.org/10.1111/rssc.12170}
}
\usage{
JRSSExamples
}
\description{
A dataset consisting of several adjacency matrices for networks taken from Parker, Ben M., Steven G. Gilmour, and John Schormans. "Optimal design of experiments on connected units with application to social networks." Journal of the Royal Statistical Society: Series C (Applied Statistics) 66.3 (2017): 455-480.).
}
\keyword{datasets}
