% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextDesign.R
\name{nextDesign}
\alias{nextDesign}
\title{Get the next design sequentially}
\usage{
nextDesign(design, k, algorithm = "sequential", par = NULL)
}
\arguments{
\item{design}{The current design}

\item{k}{the number of treatments}

\item{algorithm}{Can be "sequential" in which case all designs evaluated, or "random", in which case random designs evaluated}

\item{par}{A counting variable for the CE algorithm}
}
\description{
Calculates the next valid design in a set of length n with k possible labels
}
