% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomAdj.r
\name{randomAdj}
\alias{randomAdj}
\title{Generate Random Adjacency Matrix}
\usage{
randomAdj(n, prob)
}
\arguments{
\item{n}{number of nodes in network}

\item{prob}{probability that each link is connected}
}
\description{
This function generates an (ERdos-Renyi) random Adjacency matrix
}
\keyword{Erdos-Renyi}
\keyword{network}
